config final cross-env and config

+++ index.js
@@ -2,21 +2,25 @@ const sequelize = require("./src/config/database");
 const app = require("./src/app");
 const UserService = require("./src/user/UserService");
 
-sequelize.sync({ force: true }).then(async () => {
-  console.log("db in ready");
-  for (let i = 1; i <= 5; i++) {
-    const user = {
-      username: `user${i}`,
-      email: `user${i}@gmail.com`,
-      password: `password`,
-    };
+if (process.env.NODE_ENV === "production") {
+  sequelize.sync();
+} else {
+  sequelize.sync({ force: true }).then(async () => {
+    console.log("db in ready");
+    for (let i = 1; i <= 5; i++) {
+      const user = {
+        username: `user${i}`,
+        email: `user${i}@gmail.com`,
+        password: `password`,
+      };
 
-    await UserService.create(user);
-  }
-});
+      await UserService.create(user);
+    }
+  });
+}
 
 const PORT = 5001;
 
 app.listen(PORT, () => {
-  console.log(`app is running on port ${PORT}`);
+  console.log(`app is running on port ${PORT}, mode: ${process.env.NODE_ENV}`);
 });
